// src/interfaces/student.interface.ts
export interface Student {
  studentId?: number;
  firstname: string;
  lastname: string;
  middlename: string;
  gender: string;
  birthdate: string;
  ethnicity: string;
  degree: string;
  department: string;
  college: string;
  academicstatus: string;
  homeaddress: string;
  mobilenumber: number;
  emailaddress: string;
}